pyproject.toml
[build-system]
requires = ["uv_build>=0.10.0,<0.11.0"]
build-backend = "uv_build"

[project]
name = "sheet-ripper"
version = "0.1.4"
description = "Download and Upload Google Sheets."
authors = [{ name = "Rafay Khan", email = "contact.hazelnut560@passinbox.com" }]
license = "MIT"
license-files = ["LICEN[CS]E*"]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "google-api-python-client>=2.183.0",
  "google-auth-httplib2>=0.2.0",
  "google-auth-oauthlib>=1.2.2",
]

[project.scripts]
sheets = "sheet_ripper:__main__.main"

[project.urls]
Homepage = "https://github.com/rafaykhan-source/sheet-ripper"

[tool.rumdl.MD013]
reflow = true
